moved helpers functions in utils.py
